Move enum package location
