+++ src/main/java/uk/ac/cam/cl/dtg/segue/dos/Difficulty.java
@@ -1,0 +1,6 @@
+package uk.ac.cam.cl.dtg.segue.dos;
+
+public enum Difficulty {
+    practice_1, practice_2, practice_3, challenge_1, challenge_2, challenge_3
+    /* NOTE: lower case is necessary for correct JSON auto mapping */
+}
