+++ src/main/java/uk/ac/cam/cl/dtg/segue/dos/Stage.java
@@ -1,0 +1,6 @@
+package uk.ac.cam.cl.dtg.segue.dos;
+
+public enum Stage {
+    gcse, a_level, further_a, university, none
+    /* NOTE: lower case is necessary for correct JSON auto mapping */
+}
